requirements.txt
alembic==1.5.2
Babel==2.9.1
bcrypt==3.2.0
blinker==1.4
cffi==1.14.5
click==7.1.2
dnspython==2.1.0
email-validator==1.1.3
environ==1.0
Flask==1.1.2
Flask-Authorize==0.2.2
Flask-BabelEx==0.9.4
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.6.0
Flask-Minify==0.32
flask-paginate==0.5.0
Flask-Principal==0.4.0
Flask-Script==2.0.6
Flask-Security==3.0.0
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
gunicorn==20.1.0
htmlmin==0.1.12
humanfriendly==9.2
idna==3.2
itsdangerous==1.1.0
Jinja2==2.11.2
lesscpy==0.15.0
Mako==1.1.4
MarkupSafe==1.1.1
passlib==1.7.4
Paste==3.5.0
ply==3.11
psycopg2-binary==2.8.6
pycparser==2.20
PyJWT==2.0.1
pyreadline==2.1
python-dateutil==2.8.1
python-dotenv==0.17.1
python-editor==1.0.4
pytz==2021.1
PyYAML==5.4.1
six==1.16.0
speaklater==1.3
SQLAlchemy==1.3.22
unicode==2.8
waitress==2.0.0
Werkzeug==1.0.1
WSGIserver==1.3
WTForms==2.3.3
xxhash==2.0.2
